+++ libs/component-sample/src/lib/feature/pagination-sample/pagination-sample.component.ts
@@ -7,7 +7,15 @@ import { BehaviorSubject } from "rxjs";
 })
 export class PaginationSampleComponent implements OnInit {
 
-  public pageChange = new BehaviorSubject<object>(null);
+  page = {
+    pageNumber: 1,
+    pageSize: 25,
+    totalPages: 10
+  }
+
+  top = { id: 'top' };
+  bottom = { id: 'bottom' };
+  public pageChange = new BehaviorSubject<object>(this.page);
   constructor() { }
 
   ngOnInit() {